fix: eslint dependency cycle issue
