+++ src/helpers/strategies/ChangeStrategy.ts
@@ -1,4 +1,4 @@
-import { Payload } from "../InboxProcessor";
+import { Payload } from "@src/models/InboxItem";
 
 export interface ChangeStrategy {
   execute(payload: Payload, relId: string): Promise<void>;
