+++ src/hooks/useApp.tsx
@@ -8,7 +8,7 @@ import { checkMagicGoal, getAllLevelGoalsOfId, getGoal, updateSharedStatusOfGoal
 import { addSharedWMGoal } from "@src/api/SharedWMAPI";
 import { createDefaultGoals } from "@src/controllers/NewUserController";
 import { refreshTaskCollection } from "@src/api/TasksAPI";
-import { checkAndUpdateGoalNewUpdatesStatus, handleIncomingChanges, Payload } from "@src/helpers/InboxProcessor";
+import { checkAndUpdateGoalNewUpdatesStatus, handleIncomingChanges } from "@src/helpers/InboxProcessor";
 import { getContactSharedGoals, shareGoalWithContact } from "@src/services/contact.service";
 import { updateAllUnacceptedContacts, getContactByRelId, clearTheQueue } from "@src/api/ContactsAPI";
 import { useSetRecoilState, useRecoilValue, useRecoilState } from "recoil";
@@ -19,6 +19,7 @@ import ContactItem from "@src/models/ContactItem";
 import { SharedGoalMessage } from "@src/Interfaces/IContactMessages";
 import { checkAndCleanupDoneTodayCollection } from "@src/controllers/TaskDoneTodayController";
 import { getAllInboxItems } from "@src/api/InboxAPI";
+import { Payload } from "@src/models/InboxItem";
 
 const langFromStorage = localStorage.getItem(LocalStorageKeys.LANGUAGE)?.slice(1, -1);
 const exceptionRoutes = ["/", "/invest", "/feedback", "/donate"];
